+++ src/ui/camera.ts
@@ -1343,6 +1343,26 @@ abstract class Camera extends Evented {
             delta > 180 ? -360 :
                 delta < -180 ? 360 : 0;
     }
+
+    /**
+     * Query the current elevation of location. It return null if terrain is not enabled. the elevation is in meters relative to mean sea-level
+     * @memberof Map#
+     * @param lngLatLike [x,y] or LngLat coordinates of the location
+     * @returns {number} elevation in meters
+     */
+    queryTerrainElevation(lngLatLike: LngLatLike): number | null {
+        if (!this.terrain) {
+            return null;
+        }
+        const elevation = this.transform.getElevation(LngLat.convert(lngLatLike), this.terrain);
+        /**
+         * Different zoomlevels with different terrain-tiles the elvation-values are not the same.
+         * map.transform.elevation variable with the center-altitude.
+         * In maplibre the proj-matrix is translated by this value in negative z-direction.
+         * So we need to add this value to the elevation to get the correct value.
+         */
+        return elevation - this.transform.elevation;
+    }
 }
 
 // In debug builds, check that camera change events are fired in the correct order.